move ts file
