+++ src/svelteDirectives/isSSR.ts
@@ -1,4 +1,4 @@
-export default (function isSSR() {
+export default (function isSSR(): boolean {
     try {
         return typeof document === 'undefined';
     } catch (e) {
